sam4e8e: Move ADC code to new file afec.c
