+++ src/sam4e8e/Makefile
@@ -21,7 +21,7 @@ src-y +=  ../lib/cmsis-sam4e/gcc/system_sam4e.c \
 src-$(CONFIG_HAVE_GPIO_SPI) += sam4e8e/spi.c
 src-$(CONFIG_HAVE_GPIO_I2C) += sam4e8e/i2c.c
 src-$(CONFIG_SERIAL) += sam4e8e/serial.c generic/serial_irq.c
-src-$(CONFIG_HAVE_GPIO) += sam4e8e/gpio.c
+src-$(CONFIG_HAVE_GPIO) += sam4e8e/gpio.c sam4e8e/afec.c
 src-y += generic/crc16_ccitt.c generic/alloc.c
 src-y += generic/armcm_irq.c generic/timer_irq.c
 src-y += sam4e8e/main.c sam4e8e/timer.c
