+++ src/sam4e8e/gpio.h
@@ -31,7 +31,6 @@ struct gpio_adc {
     uint32_t chan;
 };
 
-struct gpio_adc gpio_pin_to_afec(uint8_t pin);
 struct gpio_adc gpio_adc_setup(uint8_t pin);
 uint32_t gpio_adc_sample(struct gpio_adc g);
 uint16_t gpio_adc_read(struct gpio_adc g);
